+++ app/screens/TripDashboard/Dashboard.tsx
@@ -1,11 +1,18 @@
 import { View, Text } from 'react-native'
 import React from 'react'
-import TabNavigator from '../../navigators/TabNavigator'
+import TabNavigator from '../../components/TabNavigator'
 import { Screen } from '../../components'
 
 const Dashboard = () => {
   return (
-        <TabNavigator />
+    <Screen
+        preset="fixed"
+        statusBarStyle="light"
+        backgroundImage={require("../../../assets/images/gradientBg.png")}
+        goBackHeader={true}
+        showNavBar={true}
+      >
+</Screen>
   )
 }
